+++ src/components/openGraph/OgImage.tsx
@@ -1,11 +1,11 @@
 import {Stack, Text} from '@sanity/ui'
-import React, {useMemo} from 'react'
+import {type ReactElement, useMemo} from 'react'
 import {ObjectInputProps, useFormValue} from 'sanity'
 
 import {FeedbackType} from '../../types'
 import {getOgImageValidation} from '../../utils/seoUtils'
 
-const OgImage = (props: ObjectInputProps): React.ReactElement => {
+const OgImage = (props: ObjectInputProps): ReactElement => {
   const {value, renderDefault, path} = props
 
   const seoParent = useFormValue([path[0]]) as Record<string, unknown> | null